requirements.txt
asgiref==3.4.1
Django==3.2.13
django-cors-headers==3.10.1
django-cors-middleware==1.5.0
djangorestframework==3.13.1
pkg-resources==0.0.0
pytz==2022.1
sqlparse==0.4.2
typing-extensions==4.1.1
